refactor: domain이라고 되어있는 일부 하위 패키지의 이름을 model로 변경
원래 특정 도메인이 가지는 모델 관련 하위 패키지명은 model인데 실수로 domain이라고 해놓은 것들이 있었음.
