+++ src/main/java/com/jvnlee/catchdining/domain/payment/model/PaymentType.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.domain.payment.domain;
+package com.jvnlee.catchdining.domain.payment.model;
 
 public enum PaymentType {
     CREDIT_CARD, NAVER_PAY, KAKAO_PAY
