+++ src/main/java/com/jvnlee/catchdining/domain/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.jvnlee.catchdining.domain.payment.repository;
 
-import com.jvnlee.catchdining.domain.payment.domain.Payment;
+import com.jvnlee.catchdining.domain.payment.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
